Update the procurement list when a part in use is already being ordered
Refactor PartsTableModel to PartsStoreModel

parts store dialog updates shot quantity correctly

fix clan lrm-5 procurement giving 25 shots not 24
