+++ MekHQ/src/mekhq/campaign/parts/equipment/AmmoBin.java
@@ -139,6 +139,11 @@ public class AmmoBin extends EquipmentPart implements IAcquisitionWork {
         return getFullShots() / (double) getType().getShots();
     }
 
+    @Override
+    public int getTotalQuantity() {
+        return getQuantity() * getType().getShots();
+    }
+
     public int getFullShots() {
         if (oneShot) {
             return 1;
