+++ MekHQ/src/mekhq/gui/dialog/PartsStoreDialog.java
@@ -81,7 +81,8 @@ import mekhq.campaign.parts.protomeks.ProtoMekSensor;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.work.IAcquisitionWork;
 import mekhq.gui.CampaignGUI;
-import mekhq.gui.dialog.PartsStoreDialog.PartsTableModel.PartProxy;
+import mekhq.gui.model.PartsStoreModel;
+import mekhq.gui.model.PartsStoreModel.PartProxy;
 import mekhq.gui.sorter.PartsDetailSorter;
 import mekhq.gui.utilities.JScrollPaneWithSpeed;
 
@@ -111,11 +112,10 @@ public class PartsStoreDialog extends JDialog {
 
     private final Campaign campaign;
     private final CampaignGUI campaignGUI;
-    private final PartsTableModel partsModel;
-    private TableRowSorter<PartsTableModel> partsSorter;
+    private final PartsStoreModel partsModel;
+    private TableRowSorter<PartsStoreModel> partsSorter;
     private final boolean addToCampaign;
     private Part selectedPart;
-    private Person logisticsPerson;
 
     private JTable partsTable;
     private JTextField txtFilter;
@@ -137,7 +137,7 @@ public class PartsStoreDialog extends JDialog {
         this.campaignGUI = gui;
         this.campaign = campaign;
         this.addToCampaign = add;
-        partsModel = new PartsTableModel(campaign.getPartsStore().getInventory());
+        partsModel = new PartsStoreModel(gui, campaign.getPartsStore().getInventory());
         initComponents();
         filterParts();
         setLocationRelativeTo(frame);
@@ -155,10 +155,10 @@ public class PartsStoreDialog extends JDialog {
         partsTable = new JTable(partsModel);
         partsTable.setName("partsTable");
         partsSorter = new TableRowSorter<>(partsModel);
-        partsSorter.setComparator(PartsTableModel.COL_DETAIL, new PartsDetailSorter());
+        partsSorter.setComparator(PartsStoreModel.COL_DETAIL, new PartsDetailSorter());
         partsTable.setRowSorter(partsSorter);
         TableColumn column;
-        for (int i = 0; i < PartsTableModel.N_COL; i++) {
+        for (int i = 0; i < PartsStoreModel.N_COL; i++) {
             column = partsTable.getColumnModel().getColumn(i);
             column.setPreferredWidth(partsModel.getColumnWidth(i));
             column.setCellRenderer(partsModel.getRenderer());
@@ -248,13 +248,13 @@ public class PartsStoreDialog extends JDialog {
                         addPart(true, partProxy.getPart(), 1);
                         partProxy.updateTargetAndInventories();
                         partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                              PartsTableModel.COL_TARGET);
+                              PartsStoreModel.COL_TARGET);
                         partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                              PartsTableModel.COL_TRANSIT);
+                              PartsStoreModel.COL_TRANSIT);
                         partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                              PartsTableModel.COL_SUPPLY);
+                              PartsStoreModel.COL_SUPPLY);
                         partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                              PartsTableModel.COL_QUEUE);
+                              PartsStoreModel.COL_QUEUE);
                     }
                 }
             });
@@ -282,13 +282,13 @@ public class PartsStoreDialog extends JDialog {
                             addPart(true, partProxy.getPart(), quantity);
                             partProxy.updateTargetAndInventories();
                             partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                                  PartsTableModel.COL_TARGET);
+                                  PartsStoreModel.COL_TARGET);
                             partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                                  PartsTableModel.COL_TRANSIT);
+                                  PartsStoreModel.COL_TRANSIT);
                             partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                                  PartsTableModel.COL_SUPPLY);
+                                  PartsStoreModel.COL_SUPPLY);
                             partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                                  PartsTableModel.COL_QUEUE);
+                                  PartsStoreModel.COL_QUEUE);
                         }
                     }
                 }
@@ -306,13 +306,13 @@ public class PartsStoreDialog extends JDialog {
                         addPart(false, partProxy.getPart(), 1);
                         partProxy.updateTargetAndInventories();
                         partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                              PartsTableModel.COL_TARGET);
+                              PartsStoreModel.COL_TARGET);
                         partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                              PartsTableModel.COL_TRANSIT);
+                              PartsStoreModel.COL_TRANSIT);
                         partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                              PartsTableModel.COL_SUPPLY);
+                              PartsStoreModel.COL_SUPPLY);
                         partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                              PartsTableModel.COL_QUEUE);
+                              PartsStoreModel.COL_QUEUE);
                     }
                 }
             });
@@ -343,13 +343,13 @@ public class PartsStoreDialog extends JDialog {
                             addPart(false, partProxy.getPart(), quantity);
                             partProxy.updateTargetAndInventories();
                             partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                                  PartsTableModel.COL_TARGET);
+                                  PartsStoreModel.COL_TARGET);
                             partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                                  PartsTableModel.COL_TRANSIT);
+                                  PartsStoreModel.COL_TRANSIT);
                             partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                                  PartsTableModel.COL_SUPPLY);
+                                  PartsStoreModel.COL_SUPPLY);
                             partsModel.fireTableCellUpdated(partsTable.convertRowIndexToModel(i),
-                                  PartsTableModel.COL_QUEUE);
+                                  PartsStoreModel.COL_QUEUE);
                         }
                     }
                 }
@@ -407,10 +407,10 @@ public class PartsStoreDialog extends JDialog {
 
     public void filterParts() {
         final int nGroup = choiceParts.getSelectedIndex();
-        RowFilter<PartsTableModel, Integer> partsTypeFilter = new RowFilter<>() {
+        RowFilter<PartsStoreModel, Integer> partsTypeFilter = new RowFilter<>() {
             @Override
-            public boolean include(Entry<? extends PartsTableModel, ? extends Integer> entry) {
-                PartsTableModel partsModel = entry.getModel();
+            public boolean include(Entry<? extends PartsStoreModel, ? extends Integer> entry) {
+                PartsStoreModel partsModel = entry.getModel();
                 Part part = partsModel.getPartAt(entry.getIdentifier());
 
                 if (hideImpossible.isSelected()) {
@@ -548,470 +548,5 @@ public class PartsStoreDialog extends JDialog {
             case SG_OMNI_POD -> "Empty OmniPods";
             default -> "?";
         };
-    }
-
-    private Person getLogisticsPerson() {
-        if (null == logisticsPerson) {
-            logisticsPerson = campaign.getLogisticsPerson();
-        }
-        return logisticsPerson;
-    }
-
-    /**
-     * A table model for displaying parts - similar to the one in CampaignGUI, but not exactly
-     */
-    public class PartsTableModel extends AbstractTableModel {
-        protected String[] columnNames;
-        protected ArrayList<PartProxy> data;
-
-        public final static int COL_NAME = 0;
-        public final static int COL_DETAIL = 1;
-        public final static int COL_TECH_BASE = 2;
-        public final static int COL_COST = 3;
-        public final static int COL_TON = 4;
-        public final static int COL_TARGET = 5;
-        public final static int COL_SUPPLY = 6;
-        public final static int COL_TRANSIT = 7;
-        public final static int COL_QUEUE = 8;
-        public final static int N_COL = 9;
-
-        /**
-         * Provides a lazy view to a {@link TargetRoll} for use in a UI (e.g. sorting in a table).
-         */
-        public static class TargetProxy implements Comparable<TargetProxy> {
-            private final TargetRoll target;
-            private String details;
-            private String description;
-
-            /**
-             * Creates a new proxy object for a {@link TargetRoll}.
-             *
-             * @param t The {@link TargetRoll} to be proxied. May be null.
-             */
-            public TargetProxy(@Nullable TargetRoll t) {
-                target = t;
-            }
-
-            /**
-             * Gets the target roll.
-             *
-             * @return The target roll.
-             */
-            public TargetRoll getTargetRoll() {
-                return target;
-            }
-
-            /**
-             * Gets a description of the target roll.
-             *
-             * @return A description of the target roll.
-             */
-            @Nullable
-            public String getDescription() {
-                if (null == target) {
-                    return null;
-                }
-                if (null == description) {
-                    description = target.getDesc();
-                }
-                return description;
-            }
-
-            /**
-             * Gets a string representation of a {@link TargetRoll}.
-             *
-             * @return A string representation of a {@link TargetRoll}.
-             */
-            @Override
-            public String toString() {
-                if (null == target) {
-                    return "-";
-                }
-
-                if (null == details) {
-                    details = target.getValueAsString();
-                    if (target.getValue() != TargetRoll.IMPOSSIBLE &&
-                              target.getValue() != TargetRoll.AUTOMATIC_SUCCESS &&
-                              target.getValue() != TargetRoll.AUTOMATIC_FAIL) {
-                        details += "+";
-                    }
-                }
-
-                return details;
-            }
-
-            /**
-             * Converts a {@link TargetRoll} into an integer for comparisons.
-             *
-             * @return An integer representation of the {@link TargetRoll}.
-             */
-            private int coerceTargetRoll() {
-                int r = target.getValue();
-                if (r == TargetRoll.IMPOSSIBLE) {
-                    return Integer.MAX_VALUE;
-                } else if (r == TargetRoll.AUTOMATIC_FAIL) {
-                    return Integer.MAX_VALUE - 1;
-                } else if (r == TargetRoll.AUTOMATIC_SUCCESS) {
-                    return Integer.MIN_VALUE;
-                }
-                return r;
-            }
-
-            /**
-             * {@inheritDoc}
-             *
-             * @param o The {@link TargetProxy} to compare this instance to.
-             *
-             * @return {@inheritDoc}
-             */
-            @Override
-            public int compareTo(TargetProxy o) {
-                return Integer.compare(coerceTargetRoll(), o.coerceTargetRoll());
-            }
-        }
-
-        /**
-         * Provides a container for a value formatted for display and the value itself for sorting.
-         */
-        public record FormattedValue<T extends Comparable<T>>(T value, String formatted)
-              implements Comparable<FormattedValue<T>> {
-            /**
-             * Creates a wrapper around a value and a formatted string representing the value.
-             */
-            public FormattedValue {
-            }
-
-            /**
-             * Gets the wrapped value.
-             *
-             * @return The value.
-             */
-            @Override
-            public T value() {
-                return value;
-            }
-
-            /**
-             * Gets the formatted value.
-             *
-             * @return The formatted value.
-             */
-            @Override
-            @Nonnull
-            public String toString() {
-                return formatted;
-            }
-
-            /**
-             * {@inheritDoc}
-             *
-             * @return {@inheritDoc}
-             */
-            @Override
-            public int compareTo(@Nonnull FormattedValue<T> o) {
-                return value().compareTo(o.value());
-            }
-        }
-
-        /**
-         * Provides a lazy view to a {@link Part} for use in a UI (e.g. sorting in a table).
-         */
-        public class PartProxy {
-            private final Part part;
-            private String details;
-            private TargetProxy targetProxy;
-            private FormattedValue<Money> cost;
-            private PartInventory inventories;
-            private FormattedValue<Integer> ordered;
-            private FormattedValue<Integer> supply;
-            private FormattedValue<Integer> transit;
-
-            /**
-             * Initializes a new instance of the class to provide a proxy view into a part.
-             *
-             * @param p The part to proxy. Must not be null.
-             */
-            public PartProxy(Part p) {
-                part = Objects.requireNonNull(p);
-            }
-
-            /**
-             * Updates the proxied view of the properties which changed outside the proxy.
-             */
-            public void updateTargetAndInventories() {
-                targetProxy = null;
-                inventories = null;
-                ordered = null;
-                supply = null;
-                transit = null;
-            }
-
-            /**
-             * Gets the part being proxied.
-             *
-             * @return The part being proxied.
-             */
-            public Part getPart() {
-                return part;
-            }
-
-            /**
-             * Gets the part's name.
-             *
-             * @return The part's name.
-             */
-            public String getName() {
-                return part.getName();
-            }
-
-            /**
-             * Gets the part's details.
-             *
-             * @return The part's detailed.
-             */
-            public String getDetails() {
-                if (null == details) {
-                    details = part.getDetails();
-                }
-
-                return details;
-            }
-
-            /**
-             * Gets the part's cost, suitable for use in a UI element which requires both a display value and a sortable
-             * value.
-             *
-             * @return The part's cost as a {@link FormattedValue}
-             */
-            public FormattedValue<Money> getCost() {
-                if (null == cost) {
-                    Money actualValue = part.getActualValue();
-                    cost = new FormattedValue<>(actualValue, actualValue.toAmountString());
-                }
-                return cost;
-            }
-
-            /**
-             * Gets the part's tonnage.
-             *
-             * @return The part's tonnage.
-             */
-            public double getTonnage() {
-                return Math.round(part.getTonnage() * 100) / 100.0;
-            }
-
-            /**
-             * Gets the part's tech base.
-             *
-             * @return The part's tech base.
-             */
-            public String getTechBase() {
-                return part.getTechBaseName();
-            }
-
-            /**
-             * Gets the part's {@link TargetRoll}.
-             *
-             * @return A {@link TargetProxy} representing the target roll for the part.
-             */
-            public TargetProxy getTarget() {
-                if (null == targetProxy) {
-                    IAcquisitionWork shoppingItem = part.getMissingPart();
-                    if (null == shoppingItem && part instanceof IAcquisitionWork) {
-                        shoppingItem = (IAcquisitionWork) part;
-                    }
-                    if (null != shoppingItem) {
-                        TargetRoll target = campaign.getTargetForAcquisition(shoppingItem, getLogisticsPerson(), true);
-                        targetProxy = new TargetProxy(target);
-                    } else {
-                        targetProxy = new TargetProxy(null);
-                    }
-                }
-
-                return targetProxy;
-            }
-
-            /**
-             * Gets the part's quantity on order, suitable for use in a UI element which requires both a display value
-             * and a sortable value.
-             *
-             * @return The part's quantity on order as a {@link FormattedValue}
-             */
-            public FormattedValue<Integer> getOrdered() {
-                if (null == inventories) {
-                    inventories = campaign.getPartInventory(part);
-                }
-                if (null == ordered) {
-                    ordered = new FormattedValue<>(inventories.getOrdered(), inventories.orderedAsString());
-                }
-                return ordered;
-            }
-
-            /**
-             * Gets the part's quantity on hand, suitable for use in a UI element which requires both a display value
-             * and a sortable value.
-             *
-             * @return The part's quantity on hand as a {@link FormattedValue}
-             */
-            public FormattedValue<Integer> getSupply() {
-                if (null == inventories) {
-                    inventories = campaign.getPartInventory(part);
-                }
-                if (null == supply) {
-                    supply = new FormattedValue<>(inventories.getSupply(), inventories.supplyAsString());
-                }
-                return supply;
-            }
-
-            /**
-             * Gets the part's quantity in transit, suitable for use in a UI element which requires both a display value
-             * and a sortable value.
-             *
-             * @return The part's quantity in transit as a {@link FormattedValue}
-             */
-            public FormattedValue<Integer> getTransit() {
-                if (null == inventories) {
-                    inventories = campaign.getPartInventory(part);
-                }
-                if (null == transit) {
-                    transit = new FormattedValue<>(inventories.getTransit(), inventories.transitAsString());
-                }
-                return transit;
-            }
-        }
-
-        public PartsTableModel(ArrayList<Part> inventory) {
-            data = new ArrayList<>(inventory.size());
-            for (Part part : inventory) {
-                data.add(new PartProxy(part));
-            }
-        }
-
-        @Override
-        public int getRowCount() {
-            return data.size();
-        }
-
-        @Override
-        public int getColumnCount() {
-            return N_COL;
-        }
-
-        @Override
-        public String getColumnName(int column) {
-            return switch (column) {
-                case COL_NAME -> "Name";
-                case COL_DETAIL -> "Detail";
-                case COL_COST -> "Cost";
-                case COL_TON -> "Ton";
-                case COL_TECH_BASE -> "Tech";
-                case COL_TARGET -> "Target";
-                case COL_QUEUE -> "# Ordered";
-                case COL_SUPPLY -> "# Supply";
-                case COL_TRANSIT -> "# Transit";
-                default -> "?";
-            };
-        }
-
-        @Override
-        public Object getValueAt(int row, int col) {
-            PartProxy part;
-            if (data.isEmpty()) {
-                return "";
-            } else {
-                part = data.get(row);
-            }
-            if (col == COL_NAME) {
-                return part.getName();
-            }
-            if (col == COL_DETAIL) {
-                return part.getDetails();
-            }
-            if (col == COL_COST) {
-                return part.getCost();
-            }
-            if (col == COL_TON) {
-                return part.getTonnage();
-            }
-            if (col == COL_TECH_BASE) {
-                return part.getTechBase();
-            }
-            if (col == COL_TARGET) {
-                return part.getTarget();
-            }
-            if (col == COL_SUPPLY) {
-                return part.getSupply();
-            }
-            if (col == COL_TRANSIT) {
-                return part.getTransit();
-            }
-            if (col == COL_QUEUE) {
-                return part.getOrdered();
-            }
-            return "?";
-        }
-
-        @Override
-        public Class<?> getColumnClass(int c) {
-            return getValueAt(0, c).getClass();
-        }
-
-        public PartProxy getPartProxyAt(int row) {
-            return data.get(row);
-        }
-
-        public Part getPartAt(int row) {
-            return data.get(row).getPart();
-        }
-
-        public int getColumnWidth(int c) {
-            return switch (c) {
-                case COL_NAME, COL_DETAIL -> 100;
-                case COL_COST, COL_TARGET -> 40;
-                case COL_SUPPLY, COL_TRANSIT, COL_QUEUE -> 30;
-                default -> 15;
-            };
-        }
-
-        public int getAlignment(int col) {
-            return switch (col) {
-                case COL_COST, COL_TON -> SwingConstants.RIGHT;
-                case COL_TARGET -> SwingConstants.CENTER;
-                default -> SwingConstants.LEFT;
-            };
-        }
-
-        public String getTooltip(int row, int col) {
-            PartProxy part;
-            if (data.isEmpty()) {
-                return null;
-            } else {
-                part = data.get(row);
-            }
-            if (col == COL_TARGET) {
-                return part.getTarget().getDescription();
-            }
-            return null;
-        }
-
-        public Renderer getRenderer() {
-            return new Renderer();
-        }
-
-        public class Renderer extends DefaultTableCellRenderer {
-            @Override
-            public Component getTableCellRendererComponent(JTable table, Object value, boolean isSelected,
-                  boolean hasFocus, int row, int column) {
-                super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
-                setOpaque(true);
-                int actualCol = table.convertColumnIndexToModel(column);
-                int actualRow = table.convertRowIndexToModel(row);
-                setHorizontalAlignment(getAlignment(actualCol));
-                setToolTipText(getTooltip(actualRow, actualCol));
-
-                return this;
-            }
-
-        }
     }
 }
